--- src/main/java/com/bookstore/model/LoginBody.java
@@ -1,19 +1,0 @@
-package com.bookstore.model;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public class LoginBody {
-    @NotNull
-    @NotBlank
-    private String username;
-    @NotNull
-    @NotBlank
-    private String password;
-    public String getUsername() {
-        return username;
-    }
-    public String getPassword() {
-        return password;
-    }
-}
