--- src/main/java/com/bookstore/model/RegistrationBody.java
@@ -1,52 +1,0 @@
-package com.bookstore.model;
-
-import jakarta.validation.constraints.*;
-
-public class RegistrationBody {
-    @NotNull
-    @NotBlank
-    //@Size(min = 2, max = 50)
-    private String username;
-    @NotNull
-    @NotBlank
-    //@Size(min = 6, max = 32)
-    private String password;
-    @NotNull
-    @NotBlank
-    @Email
-    private String email;
-    @NotBlank
-    @NotNull
-    private String firstName;
-    @NotBlank
-    @NotNull
-    private String lastName;
-
-    public RegistrationBody(String username, String password, String email, String firstName, String lastName) {
-        this.username = username;
-        this.password = password;
-        this.email = email;
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-}
