+++ src/main/java/com/bookstore/model/UserPrincipal.java
@@ -7,44 +7,90 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
 import java.util.Collections;
 
+/**
+ * Implementation of Spring Security's UserDetails interface that wraps the application's User entity.
+ * This class serves as an adapter between the application's user model and Spring Security's user representation,
+ * providing the core user information required by Spring Security for authentication and authorization.
+ */
 public class UserPrincipal implements UserDetails {
 
     private User user;
 
+    /**
+     * Constructs a new UserPrincipal with the specified User entity.
+     *
+     * @param user the User entity to wrap
+     */
     public UserPrincipal(User user) {
         this.user = user;
     }
 
+    /**
+     * Returns the authorities granted to the user. In this implementation,
+     * all users are granted a basic "USER" role.
+     *
+     * @return a collection containing the user's authorities
+     */
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return Collections.singleton(new SimpleGrantedAuthority("USER"));
     }
 
+    /**
+     * Returns the password used to authenticate the user.
+     *
+     * @return the user's password
+     */
     @Override
     public String getPassword() {
         return user.getPassword();
     }
 
+    /**
+     * Returns the username used to authenticate the user.
+     *
+     * @return the user's username
+     */
     @Override
     public String getUsername() {
         return user.getUsername();
     }
 
+    /**
+     * Indicates whether the user's account has expired.
+     *
+     * @return true if the user's account is valid (non-expired), false otherwise
+     */
     @Override
     public boolean isAccountNonExpired() {
         return true;
     }
 
+    /**
+     * Indicates whether the user is locked or unlocked.
+     *
+     * @return true if the user is not locked, false otherwise
+     */
     @Override
     public boolean isAccountNonLocked() {
         return true;
     }
 
+    /**
+     * Indicates whether the user's credentials (password) has expired.
+     *
+     * @return true if the user's credentials are valid (non-expired), false otherwise
+     */
     @Override
     public boolean isCredentialsNonExpired() {
         return true;
     }
 
+    /**
+     * Indicates whether the user is enabled or disabled.
+     *
+     * @return true if the user is enabled, false otherwise
+     */
     @Override
     public boolean isEnabled() {
         return true;
