+++ src/main/java/com/bookstore/service/UserService.java
@@ -1,7 +1,7 @@
 package com.bookstore.service;
 
-import com.bookstore.model.LoginBody;
-import com.bookstore.model.RegistrationBody;
+import com.bookstore.dto.LoginDTO;
+import com.bookstore.dto.RegistrationDTO;
 import com.bookstore.model.Role;
 import com.bookstore.model.User;
 import com.bookstore.repository.UserRepository;
@@ -27,7 +27,15 @@ public class UserService {
 
     private BCryptPasswordEncoder encoder = new BCryptPasswordEncoder(12);
 
-    public void register(RegistrationBody body, Role role) {
+    /**
+     * Registers a new user in the system.
+     * Creates a new user entity with the provided registration details,
+     * encrypts the password, and saves it to the database.
+     *
+     * @param body the registration information containing username, password, and user details
+     * @param role the role to be assigned to the new user
+     */
+    public void register(RegistrationDTO body, Role role) {
         User user = new User();
         user.setUsername(body.getUsername());
         user.setRole(role);
@@ -37,7 +45,13 @@ public class UserService {
         repo.save(user);
     }
 
-    public String verify(LoginBody user) {
+    /**
+     * Verifies user credentials and generates a JWT token upon successful authentication.
+     *
+     * @param user the login credentials containing username and password
+     * @return JWT token string if authentication is successful, "fail" otherwise
+     */
+    public String verify(LoginDTO user) {
         Authentication authentication = authManager.authenticate(new UsernamePasswordAuthenticationToken(user.getUsername(), user.getPassword()));
         if (authentication.isAuthenticated()) {
             return jwtService.generateToken(user.getUsername());
